+++ src/hooks/use-contact-form.ts
@@ -1,18 +1,27 @@
 'use client';
 
 import { useState } from 'react';
+import { Contact } from '@/api/api';
+import { ContactRequestModel } from '@/types/contact-types';
+import { toastError, toastSuccess } from '@/utils/toast-utils';
 
 export function useContactForm() {
-    const [fullName, setFullName] = useState('');
-    const [email, setEmail] = useState('');
+    // const [fullName, setFullName] = useState('');
+    // const [email, setEmail] = useState('');
+    const [subject, setSubject] = useState('');
     const [message, setMessage] = useState('');
+    const [isSending, setIsSending] = useState<boolean>(false);
 
-    const onFullNameChange = (event: any) => {
-        setFullName(event.target.value);
-    };
+    // const onFullNameChange = (event: any) => {
+    //     setFullName(event.target.value);
+    // };
+    //
+    // const onEmailChange = (event: any) => {
+    //     setEmail(event.target.value);
+    // };
 
-    const onEmailChange = (event: any) => {
-        setEmail(event.target.value);
+    const onSubjectChange = (event: any) => {
+        setSubject(event.target.value);
     };
 
     const onMessageChange = (event: any) => {
@@ -20,18 +29,43 @@ export function useContactForm() {
     };
 
     const resetForm = () => {
-        setFullName('');
-        setEmail('');
+        // setFullName('');
+        // setEmail('');
+        setSubject('');
         setMessage('');
     };
 
+    const onSubmit = async () => {
+        const { addContactRequestEnd } = Contact;
+        const contactRequest: ContactRequestModel = {
+            subject: 'x',
+            body: message,
+            contReqDate: new Date()
+        };
+
+        setIsSending(true);
+
+        try {
+            const res = await addContactRequestEnd(contactRequest);
+
+            resetForm();
+            toastSuccess(res);
+        } catch (err: any) {
+            toastError(err.message);
+        } finally {
+            setIsSending(false);
+        }
+    };
+
     return {
-        fullName,
-        email,
+        // fullName,
+        // email,
+        subject,
         message,
-        onFullNameChange,
-        onEmailChange,
+        // onFullNameChange,
+        // onEmailChange,
+        onSubjectChange,
         onMessageChange,
-        resetForm
+        onSubmit, resetForm, isSending
     };
 }