+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.37.0] - 2024-06-03
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Contact info list at `/contact-us`.
+- [Marek Kurańda](https://github.com/mjkuranda): Capability to post new contact request information at `/contact-us`.
 
 ## [0.36.0] - 2024-06-01
 ### Added